Add tests for describe sql
