+++ tests/unit/query_graph/interpreter/test_interpreter.py
@@ -12,64 +12,6 @@ from featurebyte.query_graph.enum import NodeOutputType, NodeType
 from featurebyte.query_graph.sql.builder import SQLOperationGraph
 from featurebyte.query_graph.sql.common import SQLType
 from featurebyte.query_graph.sql.interpreter import GraphInterpreter
-from tests.util.helper import reset_global_graph
-
-
-@pytest.fixture(name="graph", scope="function")
-def query_graph():
-    """
-    Empty query graph fixture
-    """
-    yield reset_global_graph()
-
-
-@pytest.fixture(name="node_input")
-def node_input_fixture(graph):
-    """Fixture for a generic input node"""
-    node_params = {
-        "type": "event_table",
-        "columns": ["ts", "cust_id", "a", "b"],
-        "timestamp_column": "ts",
-        "table_details": {
-            "database_name": "db",
-            "schema_name": "public",
-            "table_name": "event_table",
-        },
-        "feature_store_details": {
-            "type": "snowflake",
-            "details": {
-                "database": "db",
-                "sf_schema": "public",
-                "account": "account",
-                "warehouse": "warehouse",
-            },
-        },
-    }
-    node_input = graph.add_operation(
-        node_type=NodeType.INPUT,
-        node_params=node_params,
-        node_output_type=NodeOutputType.FRAME,
-        input_nodes=[],
-    )
-    return node_input
-
-
-@pytest.fixture(name="simple_graph", scope="function")
-def simple_graph_fixture(graph, node_input):
-    """Simple graph"""
-    proj_a = graph.add_operation(
-        node_type=NodeType.PROJECT,
-        node_params={"columns": ["a"]},
-        node_output_type=NodeOutputType.SERIES,
-        input_nodes=[node_input],
-    )
-    assign = graph.add_operation(
-        node_type=NodeType.ASSIGN,
-        node_params={"name": "a_copy"},
-        node_output_type=NodeOutputType.FRAME,
-        input_nodes=[node_input, proj_a],
-    )
-    return graph, assign
 
 
 def test_graph_interpreter_super_simple(simple_graph):
